Add the llvm.frameallocate and llvm.recoverframeallocation intrinsics
Summary:
These intrinsics allow multiple functions to share a single stack
allocation from one function's call frame. The function with the
allocation may only perform one allocation, and it must be in the entry
block.

Functions accessing the allocation call llvm.recoverframeallocation with
the function whose frame they are accessing and a frame pointer from an
active call frame of that function.

These intrinsics are very difficult to inline correctly, so the
intention is that they be introduced rarely, or at least very late
during EH preparation.

The patch is still a work in progress, but I wanted to send some code
along. It needs more comments, probably some name bikeshedding, and a
change to the inliner.

Reviewers: echristo

Subscribers: llvm-commits, chandlerc, andrew.w.kaylor

Differential Revision: http://reviews.llvm.org/D6493