[sancov] Port sancov -print-coverage-pcs to COFF
Summary:
The export table is not considered part of the object file symbol table,
so we have to look through it separately.

Also add the $arch-registered-target features that clang uses to disable
tests that require a registered backend, so that we can run the sancov
tests on Windows. LLVM's lit suite did not appear to have a per-test way
to do this, and I would rather not split up the sancov tests into
architecture directories.

Reviewers: aizatsky, kcc

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D23321