[IR] De-virtualize ~Value to save a vptr
This patch is intended as an overview of all the changes required to
devirtualize Value. I intend to split off pieces of it, such as the
devirtualization of the TerminatorInst, GlobalValue, and MemoryAccess
hierarchies, but those changes don't make sense without this context.

This change makes it invalid to directly delete a Value, User, or
Instruction pointer. Instead, such code can be rewritten to check for
nullness and call Value::deleteValue(). Value objects tend to have their
lifetimes managed through iplist, so for the most part, this is an
internal implementation detail. However, there are some places where
LLVM deletes values, and those places had to be migrated to deleteValue.
I have also created llvm::unique_value, which has a custom deleter, so
it can be used in place of std::unique_ptr<Value>.