Remove getArgumentList() in favor of arg_begin(), args(), etc
Users often call getArgumentList().size(), which is a linear way to get
the number of function arguments. arg_size(), on the other hand, is
constant time.

In general, the argument list is an implementation detail, so I've
removed it from the Funciton interface and  moved all other users to the
argument iterator APIs (arg_begin(), arg_end(), args(), arg_size()).